Breadth

Breadth.@commutative

Breadth.@distributive

Breadth.@factorisable

Breadth.@idempotent

Breadth.@involute

Breadth.@commutative — Macro

f is commutative if f(x, y) = f(y, x)

Breadth.@distributive — Macro

A function f is left distributive over g if f(x, g(y, z)) = g(f(x, y), f(x, z))

Breadth.@factorisable — Macro

A function f is left factorisable over g iff f(g(x, y), g(x, z)) = g(x, f(y, z))

Breadth.@idempotent — Macro

f is idempotent if f(x, x) = x

Breadth.@involute — Macro

A function f is invloute if f(f(x)) = x