Breadth

Breadth.ABasic

Breadth.AConstantSymbol

Breadth.AFixed

Breadth.AFunctionSymbol

Breadth.AVariable

Breadth.ABasic — Type

The type of which all objects in the class are a subtype

Breadth.AConstantSymbol — Type

roots of trees can only ever be a subtype of this can also be called singletons or atoms

Breadth.AFixed — Type

function symbols and constant symbols

Breadth.AFunctionSymbol — Type

any "known" function with arity of one or more

Breadth.AVariable — Type

unknowns like x and f all unknwons have domains (values that they can take in) and a codomain (values that it can take on)